requirements.txt
pytz
requests
python-dotenv~=0.20.0
ipython

# Django
django==4.0.7
django-timezone-field
django-cors-headers
django-debug-toolbar~=3.5.0
whitenoise

# Django 3.0 async requirements
gunicorn
uvicorn
httptools

# Databases
psycopg2-binary
dj-database-url==1.0.0

# Distributed tasks

# Telegram
python-telegram-bot==13.13


#celery
celery
django-celery-beat


#modeltranslation
django-modeltranslation

geopy~=2.2.0